Created a bot package and renamed a couple things
- I think something wasn't totally right with my python path because all
of the local file imports were broken. The best fix I could find was to
move all the bot-related code into a package.
- I also renamed basic.py to main.py because I thought it fit better but
thats just my opinion.
- I don't know if I like the name of the package, change it if you think
something else is better.
