+++ .gitignore
@@ -5,5 +5,6 @@ storage_test.toml
 runBot.bat
 
 __pycache__/
+.vscode/
 *.py[cod]
 *$py.class
