+++ bot/role_commands/color_roles.py
@@ -3,12 +3,12 @@ import logging
 import random
 from discord.ext import commands
 
-from yolamtanBot.bot import YolamtanBot
+from bot import yolamtanbot
 
 
 class ColorRoles(commands.Cog):
 
-    def __init__(self, bot: YolamtanBot):
+    def __init__(self, bot: yolamtanbot.YolamtanBot):
         self.bot = bot
