+++ main.py
@@ -1,18 +1,18 @@
-from music_player.player import Player
+from bot.music_player.player import Player
 import discord
 import logging
 from discord.ext import commands
 
 from constants import *
-from bot import YolamtanBot
-from role_commands import color_roles
-from role_commands import pronoun_roles
-from music_player import player_cog
+from bot import yolamtanbot
+from bot.role_commands import color_roles
+from bot.role_commands import pronoun_roles
+from bot.music_player import player_cog
 
 
 if __name__ == '__main__':
     # Add role commands
-    bot = YolamtanBot(data_path=DATA_FILE, env_path=ENV_FILE)
+    bot = yolamtanbot.YolamtanBot(data_path=DATA_FILE, env_path=ENV_FILE)
 
     # Set up discord.py's logger (Keep separate from my own)
     discord_logger = logging.getLogger('discord')
